requirements.txt
streamlit>=1.28.0
pandas
gspread
google-auth
tomli; python_version < "3.11"
